otasigcheck: Go back to the key-rewrite check.
Using an intermediate file had the at-the-time unintended but good
side-effect of making execution failures (absent script, no
busybox in recovery, bad selinux setup) pass gracefully. Since this
check is intended as a blacklist, blocking the update when only
sure that breakage will ensue, that was a nice thing to get.

Checking the return code, while cleaner, is making non-key-related
failures like those mentioned above abort execution of the script.
The key itself isn't even checked, yet we're aborting because of it.

So go back to the original behavior: In case we're not sure... pass.

To do:
This can be cleaned up differently, and to the same effect, by returning
an explicit error code from the script and aborting when that error
comes up. Change the script to do that and come back here. :)

Change-Id: Ic9d7c3128047c25f31e2ef416c9df4d21ca0a1a8

Revert "otasigcheck: Go back to the key-rewrite check."

sigcheck now returns an explicit abort code

This reverts commit bab150d79c7b9b1f233fc3592b69a42f0b38fbd8.

Change-Id: I2b5860ea427a4db7e29b55cc632b92c6e2910494
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>