dopush only tries to push if its arg doesn't fail
`dopush mm`, as an example, tried to do mm, ignored its result, and
went to push. Now, it does mm, and if mm exits successfully it
continues to push the results, but if mm fails dopush stops,
returning mm's return code.
This is useful for having things like:
  $ mmp && adb reboot
which now reboot the device weather the changes were pushed or not.
With this patch, the device will only get rebooted if the compilation
succeeds.

Change-Id: I001e3dd83e25a775919adbccbd49914da1e94cde
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>